+++ src/TestSuite/MockTrait.php
@@ -26,14 +26,50 @@ use PHPUnit\Framework\TestCase;
 trait MockTrait
 {
     /**
+     * Internal method to set off a test failure if a class does not exist
+     * @param string $className Class name
+     * @return void
+     * @throw \PHPUnit\Framework\AssertionFailedError
+     */
+    protected function _classExistsOrFail(string $className): void
+    {
+        class_exists($className) ?: $this->fail('Class `' . $className . '` does not exist');
+    }
+
+    /**
+     * Gets the alias name for which a test is being performed, starting from a
+     *  class or a `TestCase` class.
+     *
+     * Example: class `MyPlugin\Test\TestCase\Controller\PagesControllerTest`
+     *  will return the string `Pages`.
+     * @param string|object $class Class name as string or object
+     * @return string
+     * @since 2.19.9
+     * @throw \PHPUnit\Framework\AssertionFailedError If the class does not
+     *  exist or if it is impossible to get its alias
+     */
+    protected function getAlias($class): string
+    {
+        $class = is_object($class) ? get_class($class) : $class;
+        $this->_classExistsOrFail($class);
+
+        $alias = preg_replace('/^(\w+)(Cell|Controller|Table|Validator)(Test)?$/', '$1', get_class_short_name($class), -1, $count);
+        $count ?: $this->fail('Unable to get the alias for the `' . $class . '` class');
+
+        return $alias;
+    }
+
+    /**
      * Gets the alias for a controller
      * @param string $className Controller class name
      * @return string
-     * @throws \ReflectionException
+     * @deprecated Use instead `getAlias()`
      */
     protected function getControllerAlias(string $className): string
     {
-        return substr(get_class_short_name($className), 0, -10);
+        deprecationWarning('Deprecated. Use instead `getAlias()`');
+
+        return $this->getAlias($className);
     }
 
     /**
@@ -59,8 +95,7 @@ trait MockTrait
      */
     protected function getMockForController(string $className, ?array $methods = [], ?string $alias = null): object
     {
-        class_exists($className) ?: $this->fail('Class `' . $className . '` does not exist');
-        $alias = $alias ?: $this->getControllerAlias($className);
+        $alias = $alias ?: $this->getAlias($className);
 
         return $this->getMockBuilder($className)
             ->setConstructorArgs([null, null, $alias])
@@ -84,7 +119,7 @@ trait MockTrait
 
     /**
      * Gets the class name for which a test is being performed, starting from a
-     *  `TestCase` class
+     *  `TestCase` class.
      *
      * Example: class `MyPlugin\Test\TestCase\Controller\PagesControllerTest`
      *  will return the string `MyPlugin\Controller\PagesController`.
@@ -93,7 +128,7 @@ trait MockTrait
      *  `null` on failure
      * @since 2.18.0
      */
-    public function getOriginClassName(TestCase $testClass): ?string
+    protected function getOriginClassName(TestCase $testClass): ?string
     {
         $className = preg_replace('/^([\w\\\\]+)Test\\\\TestCase\\\\([\w\\\\]+)Test$/', '$1$2', get_class($testClass), -1, $count);
 
@@ -104,17 +139,18 @@ trait MockTrait
      * Gets the class name for which a test is being performed, starting from a
      *  `TestCase` class.
      *
-     * It fails if the class cannot be determined or it does not exist
+     * It fails if the class cannot be determined or it does not exist.
      * @param \PHPUnit\Framework\TestCase $testClass A `TestCase` class
      * @return string The class name for which a test is being performed
      * @since 2.19.2
+     * @throw \PHPUnit\Framework\AssertionFailedError If the class does not
+     *  exist or if it is impossible to get its origin class name
      */
-    public function getOriginClassNameOrFail(TestCase $testClass): string
+    protected function getOriginClassNameOrFail(TestCase $testClass): string
     {
         $className = $this->getOriginClassName($testClass);
-
         $className ?: $this->fail('Unable to get the classname for the `' . get_class($testClass) . '` class');
-        class_exists($className) ?: $this->fail('Class `' . $className . '` does not exist');
+        $this->_classExistsOrFail($className);
 
         return $className;
     }
@@ -124,12 +160,12 @@ trait MockTrait
      *  `TestCase` class.
      *
      * Example: class `MyPlugin\MySubNamespace\Test\TestCase\MyExampleTest`
-     *  will return the string `MyPlugin/MySubNamespace`.     *
+     *  will return the string `MyPlugin/MySubNamespace`.
      * @param \PHPUnit\Framework\TestCase $testClass A `TestCase` class
      * @return string
      * @since 2.19.9
      */
-    public function getPluginName(TestCase $testClass): string
+    protected function getPluginName(TestCase $testClass): string
     {
         $className = get_class($testClass);